BFS works for all three provided test problems. (Although there remains a question about how we handle the fringe and whether we test for nodes or only states)
There is also a partial implementation of the Uniform Cost Search, however, this seems to depend on our use of the path-cost, which is not yet entirely clear to me. It also depends on the blokus corners problem, which has not yet been implemented.

The blokus corners problem has been provided with an implementation of the goal state check.